Start %d support development

+++ inc/my_parse_printf.h
@@ -6,7 +6,7 @@
 /*   By: modnosum <modnosum@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/24 20:37:04 by modnosum          #+#    #+#             */
-/*   Updated: 2018/07/24 21:00:17 by modnosum         ###   ########.fr       */
+/*   Updated: 2018/07/28 17:54:07 by modnosum         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 # include <my_printf_info.h>
 
-void	my_parse_star(char const **fmt, int *number,
+void	my_parse_star(char const **fmt, size_t *number,
 		va_list *args);
 void	my_manage_size_flag(char const **fmt, t_info *info,
 		t_size_f new_size);
